pyproject.toml
[project]
dependencies = [
  "requests",
  "fastapi>=0.104.0",
  "uvicorn[standard]>=0.24.0",
  "pydantic>=2.0.0",
  "numpy>=2.3.5",
  "pillow>=12.1.1",
  "noise>=1.2.2",
  "opensimplex>=0.4.5",
  "scipy>=1.15.0",
  "matplotlib>=3.10.0",
  "pymeshlab>=2022.2",
  "open3d>=0.17.0",
  "trimesh>=3.20.0",
]
description = "Make Robotics Cute!"
license = {file = "Licenses/Apache-2.0.txt"}
name = "robocute"
readme = "README.md"
requires-python = ">=3.12,<3.14"
version = "0.2.3"

[tool.uv.workspace]
members = ["packages/*"]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true

[project.optional-dependencies]
dev = [
  "pytest>=9.0.1",
  "mypy>=1.19.0",
  "pybind11-stubgen>=2.5.5",
  "usd-core>=26.3",
  "Pillow>=10.0.0",
]
publish = [
  "mkdocs>=1.6.1",
  "mkdocs-material>=9.0.0",
  "pymdown-extensions>=10.0.0",
  "mkdocs-i18n",
]

[project.scripts]
gen = "scripts.main:generate"
pre-pack = "scripts.main:pre_pack"
prepare = "scripts.main:prepare"

[build-system]
build-backend = "uv_build"
requires = ["uv_build>=0.8.15"]

[tool.pytest.ini_options]
testpaths = [
  "test",
]
